fix: satisfy cross-platform static checks
