+++ rust/src/api/git_stage_pathspec_tests.rs
@@ -1,0 +1,32 @@
+use super::*;
+
+#[test]
+fn git_stage_treats_selected_pathspec_characters_as_literals() {
+    let repo = init_repo();
+    std::fs::write(repo.path().join("file[1].txt"), "literal bracket\n")
+        .expect("write bracket file");
+    std::fs::write(repo.path().join("file1.txt"), "glob match\n").expect("write match file");
+    run_git(repo.path(), &["add", "file[1].txt", "file1.txt"]);
+    run_git(repo.path(), &["commit", "-m", "add pathspec files"]);
+
+    std::fs::write(repo.path().join("file[1].txt"), "literal bracket changed\n")
+        .expect("modify bracket file");
+    std::fs::write(repo.path().join("file1.txt"), "glob match changed\n")
+        .expect("modify match file");
+
+    git_stage(path_str(repo.path()), Some("file[1].txt".to_string())).unwrap();
+
+    let status = git_status(path_str(repo.path())).unwrap();
+    assert!(status
+        .entries
+        .iter()
+        .any(|entry| entry.path == "file[1].txt" && entry.area == GitChangeArea::Staged));
+    assert!(status
+        .entries
+        .iter()
+        .any(|entry| entry.path == "file1.txt" && entry.area == GitChangeArea::Unstaged));
+    assert!(!status
+        .entries
+        .iter()
+        .any(|entry| entry.path == "file1.txt" && entry.area == GitChangeArea::Staged));
+}
